完成添加购物车、获取购物车列表、删除购物车、修改购物车、全选/取消全选购物车信息

+++ src/main/java/com/soft2242/shop/common/result/PageResult.java
@@ -21,11 +21,11 @@ public class PageResult<T> implements Serializable {
     @Schema(description = "列表数据")
     private List<T> list;
 
-    public PageResult(int counts, int pageSize, int page, int pages, List<T> list) {
-        this.counts = counts;
+    public PageResult(long counts, int pageSize, int page, long pages, List<T> list) {
+        this.counts = (int)counts;
         this.pageSize = pageSize;
         this.page = page;
-        this.pages = pages;
+        this.pages = (int)pages;
         this.list = list;
     }
 }
